fixed INVALID_USERNAME_ID to be long instead of int

+++ app/src/main/java/mseffner/twitchnotifier/settings/SettingsManager.java
@@ -34,7 +34,7 @@ public class SettingsManager {
     public static final int RERUN_OFFLINE = 2;
 
     // Public constant for invalid username id
-    public static final int INVALID_USERNAME_ID = -1;
+    public static final long INVALID_USERNAME_ID = -1;
 
     // Public constants to indicate what setting has changed
     public static final int SETTING_USERNAME = 0;
